% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_between_group.R
\name{dist_between_group}
\alias{dist_between_group}
\title{dist_between_group}
\usage{
dist_between_group(dna, group_df, model = "raw")
}
\arguments{
\item{dna}{DNAbin inported by ape::read.dna()}

\item{group_df}{group definition file}

\item{model}{genetic distance model available in ape::dist.dna()}
}
\value{
List consist of genetic distance among individuals and groups
}
\description{
\code{dist_within_group} Calculating genetic distances within each group
}
\seealso{
ape
}
