% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_matrix_csv.R
\name{export_matrix_csv}
\alias{export_matrix_csv}
\title{export_matrix_csv}
\usage{
export_matrix_csv(result, out_dir = "./res")
}
\arguments{
\item{result}{test}

\item{out_dir}{test}
}
\value{
List consist of genetic distance among individuals and groups
}
\description{
\code{dist_within_group} Calculating genetic distances within each group
}
\seealso{
ape
}
