% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homology_between_group.R
\name{homology_between_group}
\alias{homology_between_group}
\title{homology_between_group}
\usage{
homology_between_group(dna, group_df)
}
\arguments{
\item{dna}{DNAbin inported by ape::read.dna()}

\item{group_df}{group definition file}
}
\value{
List consist of genetic distance among individuals and groups
}
\description{
\code{dist_within_group} Calculating genetic distances within each group
}
\seealso{
ape
}
